+++ index.android.js
@@ -3,13 +3,9 @@
  * @flow
  */
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import { processColor, PointPropType, StyleSheet, View, ViewPropTypes } from 'react-native';
-import type { ViewProps } from 'react-native/Libraries/Components/View/ViewPropTypes';
-const deprecatedPropType = require('react-native/Libraries/Utilities/deprecatedPropType.js');
-const ColorPropType = require('react-native/Libraries/StyleSheet/ColorPropType.js');
+import { processColor, StyleSheet, View } from 'react-native';
 
-import NativeLinearGradient from './nativeLinearGradient';
+import NativeLinearGradient, { type Props } from './common';
 
 const convertPoint = (name, point) => {
   if (Array.isArray(point)) {
@@ -18,55 +14,24 @@ const convertPoint = (name, point) => {
       'Array type is deprecated.'
     );
   }
+  // TODO: Update Android native code to receive a {x, y} object, not an array
   if (point !== null && typeof point === 'object') {
     return [point.x, point.y];
   }
   return point;
 };
 
-type PropsType = {
-  start?: Array<number> | {x: number, y: number};
-  end?: Array<number> | {x: number, y: number};
-  colors: Array<string>;
-  locations?: Array<number>;
-  useAngle?: boolean;
-  angleCenter?: {x: number, y: number};
-  angle?: number;
-} & ViewProps;
-
 /**
  * Checks if value is a valid number. Otherwise, defaults to defaultValue.
- * 
- * @param {number} defaultValue 
+ *
+ * @param {number} defaultValue
  */
 const validNumber = (defaultValue) => (value) => {
   return typeof value === 'number' ? value : defaultValue;
 };
 
-export default class LinearGradient extends Component {
-  static propTypes = {
-    start: PropTypes.oneOfType([
-      PointPropType,
-      deprecatedPropType(
-        PropTypes.arrayOf(PropTypes.number),
-        'Use point object with {x, y} instead.'
-      )
-    ]),
-    end: PropTypes.oneOfType([
-      PointPropType,
-      deprecatedPropType(
-        PropTypes.arrayOf(PropTypes.number),
-        'Use point object with {x, y} instead.'
-      )
-    ]),
-    colors: PropTypes.arrayOf(ColorPropType).isRequired,
-    locations: PropTypes.arrayOf(PropTypes.number),
-    useAngle: PropTypes.bool,
-    angleCenter: PointPropType,
-    angle: PropTypes.number,
-    ...ViewPropTypes,
-  };
-  props: PropsType;
+export default class LinearGradient extends Component<Props> {
+  props: Props;
   gradientRef: any;
 
   static defaultProps = {
@@ -74,7 +39,7 @@ export default class LinearGradient extends Component {
     end: { x: 0.5, y: 1.0 },
   };
 
-  setNativeProps(props: PropsType) {
+  setNativeProps(props: Props) {
     this.gradientRef.setNativeProps(props);
   }
 
@@ -124,7 +89,7 @@ export default class LinearGradient extends Component {
           endPoint={convertPoint('end', end)}
           locations={locations ? locations.slice(0, colors.length) : null}
           useAngle={useAngle}
-          angleCenter={convertPoint('angleCenter', angleCenter)}
+          angleCenter={angleCenter}
           angle={angle}
           borderRadii={borderRadiiPerCorner}
         />