+++ nativeLinearGradient.js
@@ -1,4 +1,0 @@
-import { requireNativeComponent } from 'react-native';
-
-export default requireNativeComponent('BVLinearGradient', null);
-